refactor(workboard): workboard 관련 컴포넌트 폴더 구조 및 위치 변경
- 기존 구조보다 유지보수 용이하도록 수정
