+++ src/pages/wireframe-project/workboard-page/WorkBoard.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { BoardContainer, BoardWrapper } from "./WorkBoard.styled";
-import TodoColumn from "./TodoColumn";
-import InProgressColumn from "./InProgressColumn";
-import CompletedColumn from "./CompletedColumn";
+import TodoColumn from "../../../components/workboard-element/TodoColumn";
+import InProgressColumn from "../../../components/workboard-element/InProgressColumn";
+import CompletedColumn from "../../../components/workboard-element/CompletedColumn";
 
 const WorkBoard = () => {
   return (
